# docs/openapi.yaml
openapi: 3.0.0
info:
  title: InnoAI API
  version: 3.1.0
  description: |
    AI Models for Every Filipino
    
    The InnoAI platform provides access to cutting-edge AI models including chat, image generation, 
    audio processing, 3D generation, video generation, and embeddings. Built for the Philippine market 
    with local currency pricing (PHP) and payment integration.
    
    ## Authentication
    
    The API supports three authentication methods:
    
    1. **JWT Bearer Token** - For customer portal authentication
    2. **API Key** - For inference endpoints (X-API-Key header)
    3. **Admin Key** - For administrative endpoints (x-admin-key header)
    
    ## Access Control Levels
    
    | Tier | Price | Requests/Day | Tokens/Month | Concurrent | Models |
    |------|-------|--------------|--------------|------------|--------|
    | Free | ₱0 | 100 | 10K | 1 | Basic only |
    | Starter | ₱499 | 1,000 | 100K | 3 | All chat |
    | Professional | ₱1,499 | 10,000 | 1M | 10 | All models |
    | Enterprise | ₱4,999 | 100,000 | 10M | 50 | All + Priority |
    
    ## Pricing (PHP)
    
    - **Chat (Boyong Opus)**: ₱465/1M input, ₱2,325/1M output
    - **Images (FLUX Pro)**: ₱3.75/image
    - **Video (Wan 480p)**: ₱4.68/second
    - **3D Generation**: Starting at ₱0.50/model
    
    ## Base URL
    
    All API requests should be made to: `https://ai-gateway.innoserver.cloud`
    
  contact:
    name: InnovateHub Support
    email: support@innovatehub.ph
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - AdminKeyAuth: []
...
paths:
...
  /admin/analytics:
    get:
      tags:
        - Admin
      summary: Get analytics
      description: Get system analytics and usage statistics (Admin only)
      operationId: getAnalytics
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRequests:
                    type: integer
                  totalTokens:
                    type: integer
                  requestsByModel:
                    type: object
                  requestsByHour:
                    type: object
                  requestsByDay:
                    type: object
                  uptime_ms:
                    type: integer
                  uptime_human:
                    type: string
                  last24Hours:
                    type: object
                  last7Days:
                    type: object
              example:
                totalRequests: 50000
                totalTokens: 25000000
                requestsByModel:
                  inno-ai-boyong-4.5: 30000
                  flux-1.1-pro: 5000
                uptime_ms: 86400000
                uptime_human: 1d 0h 0m
                last24Hours:
                  "2026-02-21T10": 450
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/logs:
    get:
      tags:
        - Admin
      summary: Get logs
      description: Get recent request logs (Admin only)
      operationId: getLogs
      security:
        - AdminKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Number of logs to return
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LogEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/logs/{id}:
    get:
      tags:
        - Admin
      summary: Get log by ID
      description: Get a specific log entry by ID (Admin only)
      operationId: getLogById
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Log entry retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogEntry'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/reset:
    post:
      tags:
        - Admin
      summary: Reset analytics
      description: Reset all analytics data (Admin only)
      operationId: resetAnalytics
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Analytics reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: true
                message: Analytics reset
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/system:
    get:
      tags:
        - Admin
      summary: Get system info
      description: Get system information and resource usage (Admin only)
      operationId: getSystemInfo
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: System information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cpu:
                    type: string
                  memory:
                    type: string
                  disk:
                    type: string
                  nodeVersion:
                    type: string
                  platform:
                    type: string
                  pid:
                    type: integer
                  hfConfigured:
                    type: boolean
                  replicateConfigured:
                    type: boolean
                  capabilities:
                    type: object
              example:
                cpu: "25.5%"
                memory: "2048/8192 MB"
                disk: "45G/100G"
                nodeVersion: v20.11.0
                platform: linux
                pid: 12345
                hfConfigured: true
                replicateConfigured: true
                capabilities:
                  chat: true
                  images: true
                  audio: true
                  embeddings: true
                  3d: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/playground:
    post:
      tags:
        - Admin
      summary: Playground chat
      description: Test chat completions in the admin playground
      operationId: adminPlayground
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      content:
                        type: string
                model:
                  type: string
            example:
              messages:
                - role: user
                  content: Hello, how are you?
              model: inno-ai-boyong-4.5
      responses:
        '200':
          description: Playground response successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  model:
                    type: string
                  response:
                    type: string
                  usage:
                    type: object
              example:
                model: inno-ai-boyong-4.5
                response: I'm doing great! How can I help you today?
                usage:
                  prompt_tokens: 10
                  completion_tokens: 15
                  total_tokens: 25
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/api-keys:
    get:
      tags:
        - Admin
      summary: List all API keys
      description: Get all API keys (Admin only)
      operationId: adminListApiKeys
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Admin
      summary: Create admin API key
      description: Create a new API key (Admin only)
      operationId: adminCreateApiKey
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                rateLimit:
                  type: integer
                  default: 100
                monthlyLimit:
                  type: integer
                  default: 1000000
            example:
              name: Admin Test Key
              rateLimit: 1000
              monthlyLimit: 5000000
      responses:
        '200':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  key:
                    type: string
                  message:
                    type: string
              example:
                success: true
                key: inno_abc123def456...
                message: API key created. Save this key - it won't be shown again!
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/api-keys/{keyId}:
    patch:
      tags:
        - Admin
      summary: Update API key
      description: Update an API key's settings (Admin only)
      operationId: adminUpdateApiKey
      security:
        - AdminKeyAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                rateLimit:
                  type: integer
                monthlyLimit:
                  type: integer
                enabled:
                  type: boolean
      responses:
        '200':
          description: API key updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Admin
      summary: Delete API key
      description: Delete an API key (Admin only)
      operationId: adminDeleteApiKey
      security:
        - AdminKeyAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/oauth/url:
    post:
      tags:
        - Admin
      summary: Generate OAuth URL
      description: Generate OAuth URL for external provider authentication (Admin only)
      operationId: generateOAuthUrl
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                accountName:
                  type: string
            example:
              provider: antigravity
              accountName: Production Account
      responses:
        '200':
          description: OAuth URL generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  state:
                    type: string
                  expiresIn:
                    type: integer
              example:
                url: https://accounts.google.com/o/oauth2/auth?...
                state: abc123...
                expiresIn: 300
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/oauth/exchange:
    post:
      tags:
        - Admin
      summary: Exchange OAuth code
      description: Exchange OAuth authorization code for access tokens (Admin only)
      operationId: exchangeOAuthCode
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                state:
                  type: string
                callbackUrl:
                  type: string
                accountName:
                  type: string
      responses:
        '200':
          description: Tokens exchanged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  account:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/accounts:
    get:
      tags:
        - Admin
      summary: List connected accounts
      description: Get all connected OAuth accounts (Admin only)
      operationId: listAccounts
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/accounts/{id}:
    delete:
      tags:
        - Admin
      summary: Delete account
      description: Delete a connected account (Admin only)
      operationId: deleteAccount
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/accounts/{id}/refresh:
    post:
      tags:
        - Admin
      summary: Refresh account token
      description: Refresh OAuth access token for an account (Admin only)
      operationId: refreshAccountToken
      security:
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token refreshed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for customer authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for inference endpoints (format ik_xxxxxxxx...)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin key for administrative endpoints
  schemas:
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        model:
          type: string
        source:
          type: string
        apiKey:
          type: string
        promptPreview:
          type: string
        responsePreview:
          type: string
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
        latency:
          type: integer
        status:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Invalid request parameters
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Authentication required
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Resource not found
...